+++ src/components/Header/Header.js
@@ -10,10 +10,6 @@ const propTypes = {
   loginUrl: PropTypes.string,
 };
 
-const defaultProps = {
-  loginUrl: undefined,
-};
-
 function Header({ loginUrl }) {
   return (
     <div className="wkp-header">
@@ -29,6 +25,5 @@ function Header({ loginUrl }) {
 }
 
 Header.propTypes = propTypes;
-Header.defaultProps = defaultProps;
 
 export default React.memo(Header);