+++ src/components/InputIcon/InputIcon.js
@@ -33,7 +33,7 @@ const useStyles = makeStyles((theme) => {
   };
 });
 
-function InputIcon({ checked, type, "data-cy": dataCy }) {
+function InputIcon({ checked = false, type = "checkbox", "data-cy": dataCy }) {
   const classes = useStyles();
   const theme = useTheme();
   return (
@@ -68,11 +68,6 @@ InputIcon.propTypes = {
   checked: PropTypes.bool,
   type: PropTypes.oneOf(["checkbox", "radio"]),
   "data-cy": PropTypes.string,
-};
-InputIcon.defaultProps = {
-  checked: false,
-  type: "checkbox",
-  "data-cy": undefined,
 };
 
 export default InputIcon;