+++ src/components/Select/Select.js
@@ -33,10 +33,10 @@ const useStyles = makeStyles(() => {
  *  This component fits the official design of SBB,
  *  see https://angular.app.sbb.ch/angular/components/select/examples
  */
-function Select(props) {
-  const { MenuProps } = props;
+function Select({ MenuProps, ...props }) {
+  const muiMenuProps = MenuProps || {};
   const classes = useStyles();
-  const isAnchorTop = MenuProps?.anchorOrigin?.vertical === "top";
+  const isAnchorTop = muiMenuProps?.anchorOrigin?.vertical === "top";
 
   let classesMenu = {
     paper: classes.paperAnchorBottom,
@@ -62,10 +62,10 @@ function Select(props) {
           vertical: "bottom",
           horizontal: "center",
         },
-        ...MenuProps,
+        ...muiMenuProps,
         classes: {
           ...classesMenu,
-          ...(MenuProps?.classes || {}),
+          ...(muiMenuProps?.classes || {}),
         },
       }}
     />
@@ -80,10 +80,6 @@ Select.propTypes = {
       horizontal: propTypes.oneOf(["left", "right"]),
     }),
   }),
-};
-
-Select.defaultProps = {
-  MenuProps: {},
 };
 
 export default Select;