+++ src/config/topics.js
@@ -101,6 +101,7 @@ export const handicap = {
   searches: {
     Stationen: handicapStopFinder,
   },
+  minZoom: 7,
 };
 
 export const netzkarteStelen = {
@@ -125,6 +126,7 @@ export const bauprojekte = {
   projection: "EPSG:3857",
   layerInfoComponent: "ConstructionTopicInfo",
   searches: defaultSearches,
+  minZoom: 7,
 };
 
 export const infrastruktur = {
@@ -168,6 +170,7 @@ export const regionenkartePublic = {
   layers: regionenkarteLayers,
   layerInfoComponent: "RegionenkartePublicTopicInfo",
   searches: defaultSearches,
+  minZoom: 7,
 };
 
 export const tarifverbundkarte = {
@@ -176,13 +179,14 @@ export const tarifverbundkarte = {
   layerInfoComponent: "TarifverbundkarteTopicInfo",
   layers: tarifverbundkarteLayers,
   maxZoom: 12,
+  minZoom: 7,
   exportConfig: {
     getTemplateValues: () => ({
       publisher: "tobias.hauser@sbb.ch",
-      publishedAt: "12/2023",
-      dateDe: "10.12.2023",
-      dateFr: "10.12.2023",
-      year: "2023",
+      publishedAt: "12/2024",
+      dateDe: "15.12.2024",
+      dateFr: "15.12.2024",
+      year: "2024",
     }),
     getOverlayImageUrl: () => tarifverbundkarteLegend,
   },
@@ -206,6 +210,7 @@ export const zweitausbildung = {
   projection: "EPSG:3857",
   layerInfoComponent: "ZweitausbildungTopicInfo",
   searches: defaultSearches,
+  minZoom: 7,
 };
 
 export const netzentwicklung = {
@@ -217,6 +222,7 @@ export const netzentwicklung = {
   projection: "EPSG:3857",
   layerInfoComponent: "NetzentwicklungTopicInfo",
   searches: defaultSearches,
+  minZoom: 7,
 };
 
 export const beleuchtungsstaerken = {
@@ -228,6 +234,7 @@ export const beleuchtungsstaerken = {
   projection: "EPSG:3857",
   layerInfoComponent: "BeleuchtungTopicInfo",
   searches: defaultSearches,
+  minZoom: 7,
 };
 
 export const energiePublic = {
@@ -256,6 +263,7 @@ export const energiePublic = {
   projection: "EPSG:3857",
   layerInfoComponent: "EnergiePublicTopicInfo",
   searches: defaultSearches,
+  minZoom: 7,
 };
 
 export const isb = {
@@ -271,6 +279,7 @@ export const isb = {
   projection: "EPSG:3857",
   layerInfoComponent: "IsbTopicInfo",
   searches: defaultSearches,
+  minZoom: 7,
 };
 
 const sandbox = {
@@ -299,6 +308,7 @@ export const geltungsbereicheMvp = {
     shareMenu: true,
   },
   maxZoom: 14,
+  minZoom: 7,
   layers: geltungsbereicheMvpLayers,
   projection: "EPSG:3857",
   layerInfoComponent: "GeltungsbereicheTopicInfo",
@@ -460,8 +470,8 @@ export const railPlus = {
   ],
   exportConfig: {
     getTemplateValues: () => ({
-      publishedAt: "12/2023",
-      year: "2023",
+      publishedAt: "12/2024",
+      year: "2024",
     }),
     getOverlayImageUrl: () => railplusLegend,
     getExportFileName: () =>
@@ -494,7 +504,6 @@ const topics = {
     direktverbindungen,
     direktverbindungenIframe,
     zweitausbildung,
-    bauprojekte,
     handicap,
     tarifverbundkarte,
     infrastruktur,