+++ src/popups/PunctualityPopup/PunctualityPopup.js
@@ -4,6 +4,7 @@ import { useDispatch, useSelector } from "react-redux";
 import Feature from "ol/Feature";
 import RouteSchedule from "react-spatial/components/RouteSchedule";
 import TralisLayer from "../../layers/TralisLayer";
+import getDelayText, { getDelaySecure } from "../../utils/getDelayString";
 
 function PunctualityPopup({ feature, layer }) {
   const dispatch = useDispatch();
@@ -53,28 +54,31 @@ function PunctualityPopup({ feature, layer }) {
     <RouteSchedule
       trackerLayer={layer}
       lineInfos={lineInfos}
-      getDelayString={(milliseconds, isArrival) => {
-        let timeInMs = milliseconds;
-        if (timeInMs < 0) {
-          timeInMs = 0;
-        }
-        const h = Math.floor(timeInMs / 3600000);
-        const m = (isArrival ? Math.ceil : Math.floor)(
-          (timeInMs % 3600000) / 60000,
-        );
-
-        if (h === 0 && m === 0) {
-          return "+0m";
-        }
-        if (h === 0) {
-          return `+${m}m`;
-        }
-        return `+${h}h${m}m`;
-      }}
+      getDelayString={getDelayText}
       renderArrivalDelay={(delay, stop, getDelayString, getDelayColor) => {
+        let delayToDisplay = delay;
+
+        // In some edge cases when arrival and departure time are in the
+        // same minute and delay too, we need to display the departure
+        // delay instead of the arrival delay to make sure the arrival
+        // time + ceiled delay is not greater than the departure time +
+        // floored delay.
+        // see TRAFWART-1702
+        if (
+          getDelaySecure(stop.departureDelay) +
+            stop.departureTime -
+            (getDelaySecure(delay) + stop.arrivalTime) <
+          60000
+        ) {
+          delayToDisplay = getDelaySecure(stop.departureDelay);
+        }
         return (
-          <span style={{ color: getDelayColor?.(delay, stop) || "inherit" }}>
-            {`${getDelayString?.(delay, true) || ""}`}
+          <span
+            style={{
+              color: getDelayColor?.(delayToDisplay, stop) || "inherit",
+            }}
+          >
+            {`${getDelayString?.(delayToDisplay, true) || ""}`}
           </span>
         );
       }}